examples/ariadne_django/requirements.txt
neo4j-graphql-py==0.1.1
Django==3.0.8
ariadne==0.11.0